move validation fn
