+++ sdk/src/types/block/core/validation.rs
@@ -9,7 +9,7 @@ use packable::{
 };
 
 use crate::types::block::{
-    core::{verify_parents, Block, Parents},
+    core::{parent::verify_parents_sets, Block, Parents},
     protocol::{ProtocolParameters, ProtocolParametersHash},
     Error,
 };
@@ -81,7 +81,7 @@ impl ValidationBlockBuilder {
 
     /// Finishes the builder into a [`ValidationBlock`].
     pub fn finish(self) -> Result<ValidationBlock, Error> {
-        verify_parents(&self.strong_parents, &self.weak_parents, &self.shallow_like_parents)?;
+        verify_parents_sets(&self.strong_parents, &self.weak_parents, &self.shallow_like_parents)?;
 
         Ok(ValidationBlock {
             strong_parents: self.strong_parents,
@@ -172,7 +172,8 @@ impl Packable for ValidationBlock {
         let shallow_like_parents = ShallowLikeParents::unpack::<_, VERIFY>(unpacker, &())?;
 
         if VERIFY {
-            verify_parents(&strong_parents, &weak_parents, &shallow_like_parents).map_err(UnpackError::Packable)?;
+            verify_parents_sets(&strong_parents, &weak_parents, &shallow_like_parents)
+                .map_err(UnpackError::Packable)?;
         }
 
         let highest_supported_version = u8::unpack::<_, VERIFY>(unpacker, &()).coerce()?;
